pyproject.toml
[project]
name = "swing-trader"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fredapi>=0.5.2",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "yfinance>=1.3.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "mypy>=1.20.2",
    "nbconvert>=7.17.1",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["notebooks", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
exclude = ["notebooks", ".venv"]
